# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Description de mon API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "List all addresses",
                "operationId": "f59ec128002d1f767d5ea0a001040599",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Create a new address",
                "operationId": "5f527985a60b59844fa7882958f607d7",
                "requestBody": {
                    "description": "Address data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "street": {
                                        "type": "string",
                                        "example": "123 Main Street"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "City"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Country"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/addresses/{id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get a specific address by ID",
                "operationId": "f5f8e416d457384bd0b870fd03cf490b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the address",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Address not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Update an address",
                "operationId": "101f4beb7785aa47bb3707f7c27c4925",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the address",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Address data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "street": {
                                        "type": "string",
                                        "example": "Updated 123 Main Street"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Updated City"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Updated Country"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "54321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Address updated successfully"
                    },
                    "404": {
                        "description": "Address not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Delete an address",
                "operationId": "857a42b8eb7bf736b8c48d2896f0efbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the address",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address deleted successfully"
                    },
                    "404": {
                        "description": "Address not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}